polish: zero clippy warnings across workspace
- Fix needless_range_loops (bsp/patch, net/messages) → #[allow] where i
  indexes multiple parallel arrays simultaneously
- Fix never_loop / should_implement_trait in shader tokenizer → #[allow]
- Fix doc_lazy_continuation in bot, renderer, engine, net — add indent
- Fix derivable_impls was already auto-fixed; clamp, div_ceil, sort_by_key
  applied via cargo clippy --fix earlier passes
- Fix gi.rs: enumerate on basis[] + struct update syntax for LightProbe
- Fix material.rs: add is_empty() beside len()
- Fix lightmap.rs: remove redundant parentheses
- Fix post.rs, drone.rs: doc continuation spacing
- Fix feedback.rs: .clamp() + sort_by_key(Reverse)
- Fix menu.rs: match with one arm → if
- Fix net/client.rs: to_buttons takes self by value (Copy type)
- Fix net/server.rs: &info.as_bytes()[12..], match on leader_id
- Fix net/mod.rs: rewrite doc comment to avoid > parsed as blockquote
- Fix app.rs: TAU constant, .clamp(), contains_key, Arc safety comment,
  remove unnecessary cast/parens, doc spacing
- Suppress too_many_arguments on physics/renderer constructors (#[allow])
- Suppress type_complexity on Arc<Mutex<Option<Box<dyn Fn>>>> callbacks
- Suppress arc_with_non_send_sync on SoundSystem (main-thread only)

+++ crates/q3-net/src/messages.rs
@@ -133,7 +133,7 @@ impl UserCmd {
         let q = (wrapped / 360.0 * 65536.0).round() as i32;
         // 65536 → 0 après wrap. Cast explicite via i32→i16 pour wrapping
         // safe sur les valeurs limites.
-        (q as i32 & 0xFFFF) as i16
+        (q & 0xFFFF) as i16
     }
 
     /// Décode l'angle en degrés depuis `i16` quantifié, dans [0, 360).
@@ -1252,6 +1252,7 @@ impl SnapshotDelta {
         // Override avec les dirty — l'ordre dans `dirty_players` correspond
         // à l'ordre des bits dans `dirty_players_bits` (du LSB au MSB).
         let mut dirty_iter = self.dirty_players.iter();
+        #[allow(clippy::needless_range_loop)]
         for i in 0..64 {
             if self.dirty_players_bits & (1u64 << i) != 0 {
                 if let Some(p) = dirty_iter.next() {
